serving/bentoml/requirements.serving.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
backports.functools-lru-cache==1.6.5
bentoml==1.1.0
black==23.3.0
boto3==1.28.0
botocore==1.31.0
build==0.10.0
cachetools==5.3.1
cattrs==23.1.2
certifi==2023.5.7
charset-normalizer==3.1.0
chumpy==0.70
circus==0.18.0
click==8.1.3
click-option-group==0.5.6
cloudpickle==2.2.1
contextlib2==21.6.0
contourpy==1.1.0
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
deepmerge==1.1.0
Deprecated==1.2.14
docker-pycreds==0.4.0
efficientnet-pytorch==0.7.1
entrypoints==0.4
exceptiongroup==1.1.2
executing==1.2.0
fire==0.5.0
flake8==6.0.0
flake8-bugbear==23.6.5
flake8-comprehensions==3.13.0
fonttools==4.40.0
frozenlist==1.3.3
fs==2.4.16
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.21.0
google-auth-oauthlib==1.0.0
grpcio==1.56.0
h11==0.14.0
Hydra==2.5
hydra-core==1.3.2
idna==3.4
imageio==2.31.1
importlib-metadata==6.0.1
importlib-resources==5.12.0
ipykernel==6.15.0
ipython==8.12.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.1
jupyter-client==7.0.6
jupyter_core==4.12.0
kiwisolver==1.4.4
lazy_loader==0.3
libcst==1.0.1
lightning-lite==1.8.0
lightning-utilities==0.3.0
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
moreorless==0.4.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.23.1
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.5.3.56
opentelemetry-api==1.18.0
opentelemetry-instrumentation==0.39b0
opentelemetry-instrumentation-aiohttp-client==0.39b0
opentelemetry-instrumentation-asgi==0.39b0
opentelemetry-sdk==1.18.0
opentelemetry-semantic-conventions==0.39b0
opentelemetry-util-http==0.39b0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathspec==0.11.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
pip==23.1.2
pip-requirements-parser==32.0.1
pip-tools==7.1.0
platformdirs==3.8.0
plotly==5.15.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.23.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pydantic==2.1.1
pydantic_core==2.4.0
pyflakes==3.0.1
Pygments==2.15.1
pynvml==11.5.0
pyparsing==3.1.0
pyproject_hooks==1.0.0
pyrootutils==1.0.4
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytorch-lightning==1.8.0
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==19.0.2
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.4.2
rsa==4.9
s3transfer==0.6.1
schema==0.7.5
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.10.1
sentry-sdk==1.27.0
setproctitle==1.3.2
setuptools==67.8.0
simple-di==0.1.5
six==1.16.0
smmap==5.0.0
smplx==0.1.13
sniffio==1.3.0
stack-data==0.6.2
starlette==0.28.0
stdlibs==2022.10.9
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
threadpoolctl==3.2.0
tifffile==2023.7.4
toml==0.10.2
tomli==2.0.1
--extra-index-url https://download.pytorch.org/whl/cpu
torch==1.13.1+cpu
torchgeometry==0.1.2 
torchmetrics==0.11.4
torchsummary==1.5.1
--extra-index-url https://download.pytorch.org/whl/cpu
torchvision==0.14.1+cpu
tornado==6.1
tqdm==4.65.0
trailrunner==1.4.0
traitlets==5.9.0
trimesh==3.22.3
typing_extensions==4.7.1
typing-inspect==0.9.0
tzdata==2023.3
urllib3==1.26.16
usort==1.0.7
uvicorn==0.23.1
wandb==0.15.4
watchfiles==0.19.0
wcwidth==0.2.6
Werkzeug==2.3.6
wheel==0.38.4
wrapt==1.15.0
yarl==1.9.2
zipp==3.15.0
pymysql
aiobotocore